save streetview images (nearby) + test images using custom detect and test function and response with result and score + Ability to clean saved streetview images + (more fancy page for google check with ajax for check, fetch, test and all)

+++ django-backend/isac_simo/api/helpers.py
@@ -99,6 +99,7 @@ def detect_image(image_file):
     
     # IF OS Path to Image exists + IBM KEY is provided + classifier list exists
     print('Detecting Image Object...')
+    saveto = None
     if os.path.exists(file_url) and settings.IBM_API_KEY and classifier_list.detect_object_model_id:
         # Authenticate the IBM Watson API
         api_token = str(settings.IBM_API_KEY)
@@ -117,7 +118,6 @@ def detect_image(image_file):
         ext = image_file.file.url.split('.')[-1]
         filename = '{}.{}'.format(uuid.uuid4().hex, ext)
 
-        saveto = None
         if not os.path.exists(os.path.join('media/temp/')):
             saveto = os.environ.get('PROJECT_FOLDER','') + '/media/temp/'+filename
         else:
@@ -168,7 +168,8 @@ def detect_image(image_file):
         print('Object Detect False, either bad response, no index, bad format array, sorted score empty etc.')
         return False
     else:
-        os.remove(saveto)
+        if saveto:
+            os.remove(saveto)
         print('FAILED TO Detect Object - Check Token, Object Detect Model id and file existence.')
         return False
 
@@ -685,3 +686,155 @@ def object_detail():
             return content
             
     return False
+
+##################################################
+##################################################
+
+# Similar to detect_image but from temp no need to deal with image_file model (used in e.g. google map image test detect)
+def detect_temp_image(file_url):
+    # IF OS Path to Image exists + IBM KEY is provided + classifier list exists
+    print('Detecting Image Object [TEMP Google Street Images]...')
+    saveto = None
+    if os.path.exists(file_url) and settings.IBM_API_KEY and classifier_list.detect_object_model_id:
+        # Authenticate the IBM Watson API
+        api_token = str(settings.IBM_API_KEY)
+        post_data = {'collection_ids': classifier_list.detect_object_model_id, 'threshold': '0.6', 'features':'objects'}
+        auth_base = 'Basic '+str(base64.b64encode(bytes('apikey:'+api_token, 'utf-8')).decode('utf-8'))
+        print(auth_base)
+
+        post_header = {'Accept':'application/json','Authorization':auth_base}
+        
+        # Temporary Resized Image (basewidth x calcheight)(save_to_path from param comes if looped through)
+        basewidth = 300
+        temp = Image.open(file_url)
+        wpercent = (basewidth/float(temp.size[0]))
+        hsize = int((float(temp.size[1])*float(wpercent)))
+        temp = temp.resize((basewidth,hsize), Image.ANTIALIAS)
+        ext = file_url.split('.')[-1]
+        filename = '{}.{}'.format(uuid.uuid4().hex, ext)
+
+        if not os.path.exists(os.path.join('media/temp/')):
+            saveto = os.environ.get('PROJECT_FOLDER','') + '/media/temp/'+filename
+        else:
+            saveto = os.path.join('media/temp/', filename)
+
+        print(saveto)
+        
+        temp.save(saveto)
+        resized_image_open = open(saveto, 'rb')
+        
+        post_files = {
+            'images_file': resized_image_open,
+        }
+
+        # Call the API
+        response = requests.post('https://gateway.watsonplatform.net/visual-recognition/api/v4/analyze?version=2019-02-11', files=post_files, headers=post_header, data=post_data)
+        status = response.status_code
+        try:
+            content = response.json()
+        except ValueError:
+            # IBM Response is BAD
+            print('IBM Response was BAD (During Detect) - (e.g. image too large)')
+            return False
+        
+        print(status)
+        print(content)
+        # If success save the data
+        if(status == 200 or status == '200' or status == 201 or status == '201'):
+            if "collections" in content['images'][0]['objects']:
+                if(content['images'][0]['objects']['collections'][0]['objects']):
+                    sorted_by_score = sorted(content['images'][0]['objects']['collections'][0]['objects'], key=lambda k: k['score'], reverse=True)
+                    print(sorted_by_score)
+                    if(sorted_by_score and sorted_by_score[0]): # Set Score
+                        resized_image_open.close()
+                        # Return Object detected type
+                        return {
+                            'object_type': sorted_by_score[0]['object'].lower(),
+                            'file_url': file_url,
+                            'temp_image': saveto,
+                        }
+        
+        resized_image_open.close()
+        os.remove(saveto)
+        print('Object Detect False, either bad response, no index, bad format array, sorted score empty etc. [Google Street Temp]')
+        return False
+    else:
+        if saveto:
+            os.remove(saveto)
+        print('FAILED TO Detect Object - Check Token, Object Detect Model id and file existence. [Google Street Temp]')
+        return False
+
+score = None
+result = None
+### SAME AS test_image BUT FOR TEMP IMAGES (no need to deal with models and other stuffs (used for e.g. in google map images testing))
+def test_temp_images(image_file, save_to_path=None, classifier_index=0, detected_as=None):
+    global score, result
+    if not detected_as:
+        detected_as = detect_temp_image(image_file)
+    
+    if not detected_as:
+        if save_to_path:
+            os.remove(save_to_path)
+        return False
+    
+    object_type = detected_as.get('object_type')
+    image_file = detected_as.get('file_url')
+    save_to_path = detected_as.get('temp_image')
+    
+    print('Trying ' + str(classifier_index) + ' No. Classifier for ' + object_type)
+    # IF OS Path to Image exists + IBM KEY is provided + classifier list exists
+    if ( os.path.exists(save_to_path) and settings.IBM_API_KEY 
+        and classifier_index < len(classifier_list.data().get(object_type,[]))
+        and classifier_list.data().get(object_type)[classifier_index] ):
+
+        # Authenticate the IBM Watson API
+        api_token = str(settings.IBM_API_KEY)
+        classifier_ids = classifier_list.data().get(object_type)[classifier_index]
+        post_data = {'classifier_ids': classifier_ids, 'threshold': '0.6'}
+        auth_base = 'Basic '+str(base64.b64encode(bytes('apikey:'+api_token, 'utf-8')).decode('utf-8'))
+        print(auth_base)
+
+        post_header = {'Accept':'application/json','Authorization':auth_base}
+        
+        # Open the Temporarily Resized Image (save_to_path from param comes if looped through - NOTE: now comes from detected_at directly)
+        if save_to_path: # Comes from param on next recursion
+            resized_image_open = open(save_to_path, 'rb')
+        
+        post_files = {
+            'images_file': resized_image_open,
+        }
+
+        # Call the API
+        response = requests.post('https://gateway.watsonplatform.net/visual-recognition/api/v3/classify?version=2018-03-19', files=post_files, headers=post_header, data=post_data)
+        status = response.status_code
+        try:
+            content = response.json()
+        except ValueError:
+            # IBM Response is BAD
+            print('IBM Response was BAD - (e.g. image too large)')
+            return False
+        
+        print(status)
+        print(content)
+        # If success save the data
+        if(status == 200 or status == '200' or status == 201 or status == '201'):
+            if(content['images'][0]['classifiers'][0]['classes']):
+                sorted_by_score = sorted(content['images'][0]['classifiers'][0]['classes'], key=lambda k: k['score'], reverse=True)
+                print(sorted_by_score)
+                score = sorted_by_score[0]['score']
+                result = sorted_by_score[0]['class']
+                # If nogo/nogos then run with next model pipe lopping through available classifier list
+                # NOTE: later classifier_list.py shall contain the recursion_list to hold results that might require re test
+                # e.g. if recursion_list.get('wall',[]) has nogo/nogos etc. then retest it etc...
+                if sorted_by_score[0]['class'].lower() == 'nogo' or sorted_by_score[0]['class'].lower() == 'nogos':
+                    if classifier_index + 1 < len(classifier_list.data().get(object_type,[])):
+                        print('NOGOS CLASS - PASSING THROUGH NEW MODEL CLASSIFIER #'+str(classifier_index + 1))
+                        test_temp_images(image_file, save_to_path, classifier_index + 1, detected_as) #save_to_path=temp file
+
+                if(classifier_index <= 0):
+                    resized_image_open.close()
+                    os.remove(save_to_path)
+                return {'score': score, 'result': result}
+    else:
+        print('FAILED TO TEST - Check Token, Classifier ids and file existence.')
+        return False
